feat: implement pratt parsing and tokenization
This commit introduces the implementation of Pratt parsing and tokenization for the minishell.

- Added `ft_isspace.c` to `libft` to check for whitespace characters.
- Modified `Makefile` to include new source files for Pratt parsing and tokenization.
- Added `pratt_parsing.c`, `pratt_handler.c`, `extract_data.c`, and `tokenizer.c` to handle the parsing logic.
- Updated `includes/minishell.h` to include function prototypes and data structures required for the new parsing implementation.
- Modified several `libft` functions to use `size_t` instead of `unsigned int` for size and index parameters.
- Modified `srcs/main.c` and `srcs/signal.c` to use the `_data()` singleton for accessing the exit status.
- Added `ft_realloc` to `srcs/utils.c` to handle memory reallocation.

+++ libft/ft_bzero.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   ft_bzero.c                                         :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: lenygarcia <marvin@42.fr>                  +#+  +:+       +#+        */
+/*   By: macauchy <macauchy@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/03/06 18:06:23 by lenygarcia        #+#    #+#             */
-/*   Updated: 2025/05/13 11:31:10 by lengarci         ###   ########.fr       */
+/*   Updated: 2025/06/10 11:15:34 by macauchy         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 void	ft_bzero(void *s, size_t n)
 {
-	unsigned int	i;
+	size_t	i;
 	unsigned char	*d;
 
 	d = (unsigned char *)s;
